+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # GKI Salatiga+ Changelog
 
+## Alpha Development: Next Release
+
+- Info: The extended Jetpack Compose material icon pack is now enabled in Gradle
+- Fix: Element colors now follow main theme
+
 ## v0.1.5-alpha (12) --- 2024-07-27
 
 - New: Added YouTube video description display